chore: Adapt audio sender to new internal map class
