+++ src/ndi_mod.cpp
@@ -3,7 +3,6 @@
 
 #include <iostream>
 #include <map>
-#include <string.h>
 
 // lua
 #include <lauxlib.h>
@@ -14,99 +13,76 @@ extern "C" {
 // matron
 #include "lua_eval.h"
 #include "hardware/screen.h"
-// cairo
-#include <cairo.h>
 // local copies of private weaver types/methods
 #include <weaver_image.h>
 }
 
-// ndi
-#include <Processing.NDI.Lib.h>
-
-
 #define MSG(contents) \
    std::cerr << "ndi-mod: " << contents << "\n"
 
+SenderRecord::SenderRecord()
+: surface(NULL), sender(NULL), sender_name(""), buffer(NULL), buffer_size(0)
+{}
 
-class SenderRecord
+SenderRecord::SenderRecord(SenderRecord&& other) {
+    surface = other.surface;
+    sender = other.sender;
+    sender_name = other.sender_name;
+    buffer = other.buffer;
+    buffer_size = other.buffer_size;
+    frame = other.frame;
+    frame_divisor = other.frame_divisor;
+    frame_counter = other.frame_counter;
+
+    other.surface = NULL;
+    other.sender_name = "";
+    other.sender = NULL;
+    other.buffer = NULL;
+    other.buffer_size = 0;
+}
+
+SenderRecord::SenderRecord(cairo_surface_t* surface, const char* name)
+: surface(surface), sender_name(name), buffer(NULL), buffer_size(0)
 {
-public:
-    cairo_surface_t* surface;
-    NDIlib_send_instance_t sender;
-    std::string sender_name;
-    unsigned char* buffer;
-    size_t buffer_size;
-    NDIlib_video_frame_v2_t frame;
-    int frame_divisor;
-    int frame_counter;
+    NDIlib_send_create_t send_create;
+    send_create.p_ndi_name = name;
+    send_create.p_groups = NULL;
+    send_create.clock_video = false;
+    send_create.clock_audio = false;
 
-    SenderRecord()
-    : surface(NULL), sender(NULL), sender_name(""), buffer(NULL), buffer_size(0)
-    {}
-
-    SenderRecord(const SenderRecord& other) = delete;
-    SenderRecord(SenderRecord&& other) {
-        surface = other.surface;
-        sender = other.sender;
-        sender_name = other.sender_name;
-        buffer = other.buffer;
-        buffer_size = other.buffer_size;
-        frame = other.frame;
-	frame_divisor = other.frame_divisor;
-	frame_counter = other.frame_counter;
-
-        other.surface = NULL;
-        other.sender_name = "";
-        other.sender = NULL;
-        other.buffer = NULL;
-        other.buffer_size = 0;
+    sender = NDIlib_send_create(&send_create);
+    if (!sender) {
+        MSG("error creating NDI sender");
     }
 
-    SenderRecord(cairo_surface_t* surface, const char* name)
-    : surface(surface), sender_name(name), buffer(NULL), buffer_size(0)
-    {
-        NDIlib_send_create_t send_create;
-        send_create.p_ndi_name = name;
-        send_create.p_groups = NULL;
-        send_create.clock_video = false;
-        send_create.clock_audio = false;
+    frame_divisor = 1;
+    frame_counter = 0;
 
-        sender = NDIlib_send_create(&send_create);
-        if (!sender) {
-            MSG("error creating NDI sender");
-        }
+    // NDI video format: 60fps RGBA progressive (with alpha ignored.)
+    // norns cairo surfaces are CAIRO_FORMAT_ARGB32 (premultiplied ARGB.)
+    // But all four bytes are always the same, so the RGBA/ARGB mismatch
+    // doesn't matter, and we can use the surface data directly.
+    frame = {
+        .frame_rate_N = 60000,
+        .frame_rate_D = frame_divisor * 1000,
+        .FourCC = NDIlib_FourCC_type_RGBX,
+        .frame_format_type = NDIlib_frame_format_type_progressive
+    };
 
-	frame_divisor = 1;
-	frame_counter = 0;
+    cairo_surface_reference(surface); // increase refcount
+    MSG("NDI sender \"" << sender_name << "\" created");
+}
 
-        // NDI video format: 60fps RGBA progressive (with alpha ignored.)
-        // norns cairo surfaces are CAIRO_FORMAT_ARGB32 (premultiplied ARGB.)
-        // But all four bytes are always the same, so the RGBA/ARGB mismatch
-        // doesn't matter, and we can use the surface data directly.
-        frame = {
-            .frame_rate_N = 60000,
-            .frame_rate_D = frame_divisor * 1000,
-            .FourCC = NDIlib_FourCC_type_RGBX,
-            .frame_format_type = NDIlib_frame_format_type_progressive
-        };
-
-        cairo_surface_reference(surface); // increase refcount
-        MSG("NDI sender \"" << sender_name << "\" created");
+SenderRecord::~SenderRecord() {
+    if (sender != NULL) {
+        NDIlib_send_destroy(sender);
+        free(buffer);
+        MSG("NDI sender destroyed");
     }
-
-    ~SenderRecord() {
-        if (sender != NULL) {
-            NDIlib_send_destroy(sender);
-            free(buffer);
-            MSG("NDI sender destroyed");
-        }
-        if (surface != NULL) {
-            cairo_surface_destroy(surface); // decrease refcount
-        }
+    if (surface != NULL) {
+        cairo_surface_destroy(surface); // decrease refcount
     }
-
-    void send();
-};
+}
 
 std::map<cairo_surface_t*, SenderRecord> surface_sender_map;
